Restructured project folders
