+++ src/Nox.Cli/Actions/NoxWorkflowContext.cs
@@ -1,5 +1,5 @@
 ﻿using Microsoft.Extensions.Configuration;
-using Nox.Cli.Actions.Configuration;
+using Nox.Cli.Configuration;
 using Nox.Core.Configuration;
 using Nox.Core.Interfaces.Configuration;
 using Spectre.Console;
