+++ src/Nox.Cli/Commands/DynamicCommand.cs
@@ -3,7 +3,7 @@
 using Helpers;
 using Microsoft.Extensions.Configuration;
 using Nox.Cli.Actions;
-using Nox.Cli.Actions.Configuration;
+using Nox.Cli.Configuration;
 using Nox.Core.Interfaces.Configuration;
 using Spectre.Console;
 using Spectre.Console.Cli;
