+++ src/Nox.Cli/Configuration/ActionConfiguration.cs
@@ -1,5 +1,6 @@
-﻿
-namespace Nox.Cli.Actions.Configuration;
+﻿using Nox.Cli.Actions;
+
+namespace Nox.Cli.Configuration;
 
 public class ActionConfiguration
 {
@@ -8,8 +9,8 @@ public class ActionConfiguration
     public string Job { get; set; } = string.Empty;
     public string Name { get; set; } = string.Empty;
     public string Uses { get; set; } = string.Empty;
-    public Dictionary<string,object> With { get; set; } = new ();
-    public Dictionary<string,string>? Validate { get; set; }
+    public Dictionary<string, object> With { get; set; } = new();
+    public Dictionary<string, string>? Validate { get; set; }
     public NoxActionDisplayMessage? Display { get; set; }
     public bool ContinueOnError { get; set; } = false;
 }
