+++ src/Nox.Cli/Configuration/CliConfiguration.cs
@@ -1,5 +1,4 @@
-﻿
-namespace Nox.Cli.Actions.Configuration;
+﻿namespace Nox.Cli.Configuration;
 
 public class CliConfiguration
 {
