+++ src/Nox.Cli/Configuration/ManifestConfiguration.cs
@@ -1,9 +1,9 @@
-﻿namespace Nox.Cli.Actions.Configuration;
+﻿namespace Nox.Cli.Configuration;
 
 public class ManifestConfiguration
 {
     public SecretsConfiguration Secrets { get; set; } = new();
 
     public List<BranchesConfiguration>? Branches { get; set; } = null;
-    
+
 }
