+++ src/Nox.Cli/Configuration/StepConfiguration.cs
@@ -1,5 +1,4 @@
-﻿
-namespace Nox.Cli.Actions.Configuration;
+﻿namespace Nox.Cli.Configuration;
 
 public class StepConfiguration
 {
