+++ src/Nox.Cli/Configuration/WorkflowConfiguration.cs
@@ -1,10 +1,9 @@
-﻿
-namespace Nox.Cli.Actions.Configuration;
+﻿namespace Nox.Cli.Configuration;
 
 public class WorkflowConfiguration
 {
     public string Name { get; set; } = string.Empty;
     public CliConfiguration Cli { get; set; } = new();
-    public Dictionary<string, StepConfiguration> Jobs { get; set; } = new ();
+    public Dictionary<string, StepConfiguration> Jobs { get; set; } = new();
 }
