+++ src/Nox.Cli/Extensions/ConfiguratorExtensions.cs
@@ -9,7 +9,7 @@ using YamlDotNet.Serialization.NamingConventions;
 using System.Text.Json;
 using Spectre.Console;
 using Nox.Cli.Actions.Configuration;
-using System.ComponentModel;
+using Nox.Cli.Services.Caching;
 
 namespace Nox.Cli;
 
@@ -129,7 +129,9 @@ internal static class ConfiguratorExtensions
     private static void GetOnlineWorkflows(Dictionary<string, string> yamlFiles, string tid, string cachePath, NoxCliCache cache)
     {
         var client = new RestClient($"https://noxorg.dev/workflows/{tid}/");
-        var request = new RestRequest("scripts") { Method = Method.Get };
+
+        var request = new RestRequest("files") { Method = Method.Get };
+
         request.AddHeader("Accept", "application/json");
 
         // Get list of files on server
@@ -165,7 +167,7 @@ internal static class ConfiguratorExtensions
 
                 yaml = client.Execute(request).Content;
 
-                if (yaml == null) throw new Exception($"Couldn't download script {file.Name}");
+                if (yaml == null) throw new Exception($"Couldn't download workflow {file.Name}");
 
                 File.WriteAllText(Path.Combine(workflowCachePath, file.Name), yaml);
             }
@@ -273,106 +275,5 @@ internal static class ConfiguratorExtensions
 
         return files.Select(f => f.FullName).ToArray();
     }
-}
-
-
-public class NoxCliCache : IChangeTracking
-{
-    private string _upn = null!;
-    public string Upn
-    {
-        get => _upn;
-        set
-        {
-            if (_upn != value)
-            {
-                _upn = value;
-                IsChanged = true;
-            }
-        }
-    }
-
-    private string _tid = null!;
-    public string Tid
-    {
-        get => _tid;
-        set
-        {
-            if (_tid != value)
-            {
-                _tid = value;
-                IsChanged = true;
-            }
-        }
-    }
-
-    private DateTimeOffset _expires;
-    public DateTimeOffset Expires
-    {
-        get => _expires;
-        set
-        {
-            if (_expires != value)
-            {
-                _expires = value;
-                IsChanged = true;
-            }
-        }
-    }
-
-    private List<RemoteFileInfo> _fileInfo = new();
-    public List<RemoteFileInfo> FileInfo
-    {
-        get => _fileInfo;
-        set
-        {
-            var intersect = _fileInfo.Intersect(value);
-            var count = intersect.ToList().Count;
-            if (count != value.Count || count != _fileInfo.Count)
-            {
-                _fileInfo = value;
-                IsChanged = true;
-            }
-        }
-    }
-
-    public string CacheFile { get; private set; } = string.Empty;
-
-    public bool IsChanged { get; private set; }
-    public void AcceptChanges() => IsChanged = false;
-    public NoxCliCache()  {}
-
-    public NoxCliCache(string cacheFile)
-    {
-        CacheFile = cacheFile;
-    }
-
-    public void Save()
-    {
-        if (IsChanged)
-        {
-            File.WriteAllText(CacheFile, JsonSerializer.Serialize(this));
-            IsChanged = false;
-        }
-    }
-
-    public static NoxCliCache Load(string cacheFile)
-    {
-        var cache = JsonSerializer.Deserialize<NoxCliCache>(File.ReadAllText(cacheFile))!;
-        cache.CacheFile = cacheFile;
-        cache.IsChanged = false;
-        return cache;
-    }
-
-}
-
-public class RemoteFileInfo : IEquatable<RemoteFileInfo>
-{
-    public string Name { get; set; } = string.Empty;
-    public int Size { get; set; } = 0;
-    public string ShaChecksum { get; set; } = string.Empty;
-
-    public bool Equals(RemoteFileInfo? other) => ShaChecksum == other?.ShaChecksum;
-    public override int GetHashCode() => (ShaChecksum).GetHashCode();
 }
